moved AddItemModal to its own folder
